Calculating checkout basket amount
On this change I'm implementing the calculation strategies of
the amount of a checkout basket.

For now we have two kinds of discount. The first is `BuyTwoPayOne`
that when a customer buy two or more equal products, he don't need
to pay for one of them. The second is `BulkDiscount` that when a
customer buy more than a quantity of the same product, the price
have a decrease on all of them.

The products with this discounts can be configured on the `main`
file.